+++ plugins/client-js/importMapPlugin.js
@@ -11,10 +11,11 @@ const importMap = {
   'page-note': './js/PageNoteElement.js',
   'npm:fix-title-attr': 'https://esm.sh/fix-title-attr@0.9.5',
   'fix-title-attr': `${githubPages}/fix-title-attr/index.js`,
-  'share-button': `${elementServer}/MySharingWidgetElement.js`,
+  'ndf-elements/dia': `${githubPages}/elements/src/dia.js`,
+  /* 'share-button': `${elementServer}/MySharingWidgetElement.js`,
   'site-search': `${elementServer}/MySearchElement.js`,
   'search-api': `${elementServer}/MySearchApiElement.js`,
-  'site-counter': `${elementServer}/MySiteCounterElement.js`,
+  'site-counter': `${elementServer}/MySiteCounterElement.js`, */
   'validator-button': `${elementServer}/MyTestElement.js`
 };