+++ plugins/client-js/js/app.js
@@ -1,15 +1,15 @@
 import PageNoteElement from 'page-note';
-import ShareButtonElement from 'share-button';
-import SearchApiElement from 'search-api';
-// import SiteSearchElement from 'site-search';
-import fixTitleAttributes, { fixArchiveLinks } from 'fix-title-attr';
+import { MySearchApiElement, MySharingWidgetElement, MySiteCounterElement } from 'ndf-elements/dia';
+import { fixTitleAttributes, fixArchiveLinks, TitleTipElement } from 'fix-title-attr';
 
 const { customElements } = window;
 
 customElements.define('page-note', PageNoteElement);
-customElements.define('share-button', ShareButtonElement);
-customElements.define('search-api', SearchApiElement);
+customElements.define('share-button', MySharingWidgetElement);
+customElements.define('search-api', MySearchApiElement);
+customElements.define('site-counter', MySiteCounterElement);
 // customElements.define('site-search', SiteSearchElement);
+customElements.define('title-tip', TitleTipElement);
 
 // Only add the HTML validator button when serving locally.
 if (pageInfo().isServeMode) {
@@ -18,10 +18,8 @@ if (pageInfo().isServeMode) {
   });
 }
 
-fixTitleAttributes({
-  titleSelector: 'main [ title ]'
-});
 fixArchiveLinks();
+fixTitleAttributes();
 
 document.documentElement.classList.remove('no-js');
 document.documentElement.classList.add('js');